Move SockJS tests to io.vertx.ext.web.handler.sockjs to mimic src package structure
