+++ vertx-web/src/test/java/io/vertx/ext/web/handler/sockjs/SockJSProtocolTest.java
@@ -13,12 +13,12 @@
  *
  *  You may elect to redistribute this code under either of these licenses.
  */
-package io.vertx.ext.web.handler;
+package io.vertx.ext.web.handler.sockjs;
 
 import io.vertx.core.logging.Logger;
 import io.vertx.core.logging.LoggerFactory;
 import io.vertx.ext.web.WebTestBase;
-import io.vertx.ext.web.handler.sockjs.SockJSHandler;
+
 import org.junit.Test;
 
 import java.io.BufferedReader;
